Refactor unwrap_cluster into it's own class
